alpha version using non-current safe patch
