+++ src/api/general-services/pipeline-manage/constructors/get-unassigned-pod.js
@@ -5,15 +5,10 @@ const getUnassignedPod = (context, step) => {
     clusterInfo,
   } = context;
 
-
-  // const MAX_HELMRELEASE_NAME_LENGTH = 53;
-  // const releaseName = `${processName}-${experimentId}`
-  // .substring(0, MAX_HELMRELEASE_NAME_LENGTH - 1);
-
   return {
     ...step,
     Type: 'Task',
-    Comment: 'Attempts to create a Kubernetes Job+Service for the pipeline runner.',
+    Comment: 'Retrieves an unassigned pipeline pod.',
     Resource: 'arn:aws:states:::eks:call',
     Parameters: {
       ClusterName: clusterInfo.name,
@@ -23,44 +18,9 @@ const getUnassignedPod = (context, step) => {
       Path: `/api/v1/namespaces/${config.pipelineNamespace}/pods`,
       QueryParameters: {
         labelSelector: [
-          '!activityArn',
+          '!activityId,type=pipeline',
         ],
       },
-      // Path: `/apis/helm.fluxcd.io/v1/namespaces/${config.pipelineNamespace}/helmreleases`,
-      // RequestBody: {
-      //   apiVersion: 'helm.fluxcd.io/v1',
-      //   kind: 'HelmRelease',
-      //   metadata: {
-      //     name: releaseName,
-      //     namespace: config.pipelineNamespace,
-      //     annotations: {
-      //       'fluxcd.io/automated': 'true',
-      //     },
-      //     labels: {
-      //       sandboxId: config.sandboxId,
-      //       type: 'pipeline',
-      //       experimentId,
-      //     },
-      //   },
-      //   spec: {
-      //     releaseName,
-      //     chart: {
-      //       git: 'git@github.com:biomage-ltd/pipeline',
-      //       path: 'pipeline-runner/chart',
-      //       ref: pipelineArtifacts.chartRef,
-      //     },
-      //     values: {
-      //       experimentId,
-      //       image: pipelineArtifacts.pipelineRunner,
-      //       namespace: config.pipelineNamespace,
-      //       sandboxId: config.sandboxId,
-      //       awsAccountId: accountId,
-      //       clusterEnv: config.clusterEnv,
-      //       awsRegion: config.awsRegion,
-      //       activityArn,
-      //     },
-      //   },
-      // },
     },
     Retry: [
       {
